switched default filesystem to ext4 from btrfs
Although I consider BTRFS to be a superior filesystem, it does not
support swap files. Swap files are often used in Vagrant when
additional memory is not available on the host. This remedies the
situation uitl either swapfiles are supported on BTRFS, or there
is no real use case for swapfiles in vagrant boxes.